Introduced BigDecimal values instead of ints in WeeklyStats; Refactored StateLogServiceImpl accordingly; Completed StateLogServiceImplTest.

+++ emo-check/src/main/java/com/github/vbabin/emocheck/state/WeeklyStats.java
@@ -2,9 +2,11 @@ package com.github.vbabin.emocheck.state;
 
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 public class WeeklyStats {
-    int badState;
-    int goodState;
-    int excellentState;
+    BigDecimal badState;
+    BigDecimal goodState;
+    BigDecimal excellentState;
 }
